refactor(core/search): Remove duplicate *SearchResult interfaces
