+++ app/scripts/modules/core/src/search/searchResult/index.ts
@@ -1,4 +1,3 @@
 export * from './AbstractBaseResultRenderer';
 export * from './searchResultsType.registry';
-export * from './model';
 export * from './renderer/DefaultSearchResultsRenderer';
