--- app/scripts/modules/core/src/search/searchResult/model/IClusterSearchResult.ts
@@ -1,9 +1,0 @@
-import { ISearchResult } from 'core/search';
-
-export interface IClusterSearchResult extends ISearchResult {
-  account: string;
-  application: string;
-  cluster: string;
-  email?: string;
-  stack: string;
-}
