--- app/scripts/modules/core/src/search/searchResult/model/IInstanceSearchResult.ts
@@ -1,12 +1,0 @@
-export interface IInstanceSearchResult {
-  account: string;
-  application: string;
-  cluster: string;
-  displayName: string;
-  href: string;
-  instanceId: string;
-  provider: string;
-  region: string;
-  serverGroup: string;
-  type: string;
-}
