--- app/scripts/modules/core/src/search/searchResult/model/ILoadBalancerSearchResult.ts
@@ -1,15 +1,0 @@
-export interface ILoadBalancerSearchResult {
-  account: string;
-  application: string;
-  detail: string;
-  displayName: string;
-  href: string;
-  loadBalancer: string;
-  loadBalancerType: string;
-  provider: string;
-  region: string;
-  stack: string;
-  type: string;
-  url: string;
-  vpcId: string;
-}
