--- app/scripts/modules/core/src/search/searchResult/model/IProjectSearchResult.ts
@@ -1,18 +1,0 @@
-import { IProjectConfig } from 'core/domain';
-
-export interface IProjectSearchResult {
-  config: IProjectConfig;
-  applications: string[];
-  clusters: string[];
-  pipelineConfigId: string;
-  createTs: number;
-  displayName: string;
-  email: string;
-  href: string;
-  id: string;
-  lastModifiedBy: string;
-  name: string;
-  type: string;
-  updateTs: number;
-  url: string;
-}
