--- app/scripts/modules/core/src/search/searchResult/model/ISecurityGroupSearchResult.ts
@@ -1,12 +1,0 @@
-export interface ISecurityGroupSearchResult {
-  account: string;
-  application: string;
-  href: string;
-  id: string;
-  name: string;
-  provider: string;
-  region: string;
-  type: string;
-  url: string;
-  vpcId: string;
-}
