--- app/scripts/modules/core/src/search/searchResult/model/IServerGroupSearchResult.ts
@@ -1,14 +1,0 @@
-import { ISearchResult } from 'core/search';
-
-export interface IServerGroupSearchResult extends ISearchResult {
-  account: string;
-  application: string;
-  cluster: string;
-  detail: string;
-  email?: string;
-  region: string;
-  sequence: string;
-  serverGroup: string;
-  stack: string;
-  url: string;
-}
