--- app/scripts/modules/core/src/search/searchResult/model/index.ts
@@ -1,7 +1,0 @@
-export * from './IApplicationSearchResult';
-export * from './IClusterSearchResult';
-export * from './IInstanceSearchResult';
-export * from './ILoadBalancerSearchResult';
-export * from './IProjectSearchResult';
-export * from './ISecurityGroupSearchResult';
-export * from './IServerGroupSearchResult';
